+++ vector/src/main/java/im/vector/app/features/home/room/detail/RoomDetailAction.kt
@@ -88,5 +88,6 @@ sealed class RoomDetailAction : VectorViewModelAction {
                                          val userJustAccepted: Boolean,
                                          val grantedEvents: RoomDetailViewEvents) : RoomDetailAction()
 
+    data class OpenOrCreateDm(val userId: String) : RoomDetailAction()
     data class JumpToReadReceipt(val userId: String) : RoomDetailAction()
 }